requirements.txt
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-cloud-secret-manager==2.15.1
gunicorn==20.1.0
pandas==1.5.3
psycopg2-binary==2.9.5